+++ connectors/src/connectors/confluence/temporal/activities.ts
@@ -1,41 +1,41 @@
 import { assertNever } from "@dust-tt/client";
-import { Op } from "sequelize";
-import TurndownService from "turndown";
 
-import type {
-  ConfluenceEntityRef,
-  ConfluenceFolderRef,
-  ConfluencePageRef,
-} from "@connectors/connectors/confluence/lib/confluence_api";
+import type { ConfluenceEntityRef } from "@connectors/connectors/confluence/lib/confluence_api";
 import {
   bulkFetchConfluencePageRefs,
   getActiveChildEntityRefs,
   pageHasReadRestrictions,
 } from "@connectors/connectors/confluence/lib/confluence_api";
-import type { ConfluencePageWithBodyType } from "@connectors/connectors/confluence/lib/confluence_client";
-import { ConfluenceClient } from "@connectors/connectors/confluence/lib/confluence_client";
 import {
-  getConfluenceEntityParentIds,
+  confluenceCheckAndUpsertSingleFolder,
+  confluenceRemoveAllFoldersInSpace,
+  confluenceRemoveUnvisitedFolders,
+} from "@connectors/connectors/confluence/lib/content/folders";
+import {
+  confluenceCheckAndUpsertSinglePage,
+  confluenceRemoveAllPagesInSpace,
+  confluenceRemoveUnvisitedPages,
+  confluenceUpsertPageToDataSource,
+  upsertConfluencePageInDb,
+} from "@connectors/connectors/confluence/lib/content/pages";
+import {
+  getConfluenceContentParentIds,
   getSpaceHierarchy,
 } from "@connectors/connectors/confluence/lib/hierarchy";
 import {
-  makeEntityInternalId,
   makeFolderInternalId,
   makePageInternalId,
   makeSpaceInternalId,
 } from "@connectors/connectors/confluence/lib/internal_ids";
-import { makeConfluenceDocumentUrl } from "@connectors/connectors/confluence/temporal/workflow_ids";
-import { filterCustomTags } from "@connectors/connectors/shared/tags";
+import {
+  getConfluenceClient,
+  getConfluenceConfig,
+} from "@connectors/connectors/confluence/lib/utils";
 import { dataSourceConfigFromConnector } from "@connectors/lib/api/data_source_config";
 import { concurrentExecutor } from "@connectors/lib/async_utils";
-import type { UpsertDataSourceDocumentParams } from "@connectors/lib/data_sources";
 import {
-  deleteDataSourceDocument,
   deleteDataSourceFolder,
-  renderDocumentTitleAndContent,
-  renderMarkdownSection,
   updateDataSourceDocumentParents,
-  upsertDataSourceDocument,
   upsertDataSourceFolder,
 } from "@connectors/lib/data_sources";
 import {
@@ -48,23 +48,16 @@ import {
   ConfluencePage,
   ConfluenceSpace,
 } from "@connectors/lib/models/confluence";
-import { getOAuthConnectionAccessTokenWithThrow } from "@connectors/lib/oauth";
 import { syncStarted, syncSucceeded } from "@connectors/lib/sync_status";
 import { heartbeat } from "@connectors/lib/temporal";
 import mainLogger from "@connectors/logger/logger";
 import { ConnectorResource } from "@connectors/resources/connector_resource";
-import type { DataSourceConfig, ModelId } from "@connectors/types";
+import type { ModelId } from "@connectors/types";
 import {
   ConfluenceClientError,
   INTERNAL_MIME_TYPES,
   isConfluenceNotFoundError,
 } from "@connectors/types";
-
-/**
- * This type represents the ID that should be passed as parentId to a content node to hide it from the UI.
- * This behavior is typically used to hide content nodes whose position in the ContentNodeTree cannot be resolved at time of upsertion.
- */
-export const HiddenContentNodeParentId = "__hidden_syncing_content__";
 
 const UPSERT_CONCURRENT_LIMIT = 10;
 
@@ -78,8 +71,6 @@ const logger = mainLogger.child({
   provider: "confluence",
 });
 
-const turndownService = new TurndownService();
-
 async function fetchConfluenceConnector(connectorId: ModelId) {
   const connector = await ConnectorResource.fetchById(connectorId);
   if (!connector) {
@@ -87,52 +78,6 @@ async function fetchConfluenceConnector(connectorId: ModelId) {
   }
 
   return connector;
-}
-
-async function getConfluenceAccessTokenWithThrow(connectionId: string) {
-  const token = await getOAuthConnectionAccessTokenWithThrow({
-    logger,
-    provider: "confluence",
-    connectionId,
-  });
-
-  return token.access_token;
-}
-
-export async function getConfluenceClient(config: {
-  cloudId?: string;
-  connectorId: ModelId;
-}): Promise<ConfluenceClient>;
-export async function getConfluenceClient(
-  config: { cloudId?: string },
-  connector: ConnectorResource
-): Promise<ConfluenceClient>;
-export async function getConfluenceClient(
-  config: {
-    cloudId?: string;
-    connectorId?: ModelId;
-  },
-  connector?: ConnectorResource
-) {
-  const { cloudId, connectorId } = config;
-
-  // Ensure the connector is fetched if not directly provided.
-  const effectiveConnector =
-    connector ??
-    (connectorId ? await fetchConfluenceConnector(connectorId) : undefined);
-
-  if (!effectiveConnector) {
-    throw new Error("A valid connector or connectorId must be provided.");
-  }
-
-  const accessToken = await getConfluenceAccessTokenWithThrow(
-    effectiveConnector.connectionId
-  );
-
-  return new ConfluenceClient(accessToken, {
-    cloudId,
-    useProxy: effectiveConnector.useProxy ?? false,
-  });
 }
 
 export async function getSpaceIdsToSyncActivity(connectorId: ModelId) {
@@ -149,18 +94,7 @@ export async function getSpaceIdsToSyncActivity(connectorId: ModelId) {
 export async function fetchConfluenceConfigurationActivity(
   connectorId: ModelId
 ) {
-  const confluenceConfig = await ConfluenceConfiguration.findOne({
-    where: {
-      connectorId,
-    },
-  });
-  if (!confluenceConfig) {
-    throw new Error(
-      `Confluence configuration not found (connectorId: ${connectorId})`
-    );
-  }
-
-  return confluenceConfig;
+  return getConfluenceConfig({ connectorId });
 }
 
 export async function confluenceSaveStartSyncActivity(connectorId: ModelId) {
@@ -262,162 +196,6 @@ export async function confluenceUpsertSpaceFolderActivity({
   }
 }
 
-export async function markPageHasVisited({
-  connectorId,
-  pageId,
-  spaceId,
-  visitedAtMs,
-}: {
-  connectorId: ModelId;
-  pageId: string;
-  spaceId: string;
-  visitedAtMs: number;
-}) {
-  await ConfluencePage.update(
-    {
-      lastVisitedAt: new Date(visitedAtMs),
-    },
-    {
-      where: {
-        connectorId,
-        pageId,
-        spaceId,
-      },
-    }
-  );
-}
-
-export async function markFolderHasVisited({
-  connectorId,
-  folderId,
-  spaceId,
-  visitedAtMs,
-}: {
-  connectorId: ModelId;
-  folderId: string;
-  spaceId: string;
-  visitedAtMs: number;
-}) {
-  await ConfluenceFolder.update(
-    {
-      lastVisitedAt: new Date(visitedAtMs),
-    },
-    {
-      where: {
-        connectorId,
-        folderId,
-        spaceId,
-      },
-    }
-  );
-}
-
-interface ConfluenceUpsertPageInput {
-  page: NonNullable<Awaited<ReturnType<ConfluenceClient["getPageById"]>>>;
-  spaceName: string;
-  parents: [string, string, ...string[]];
-  confluenceConfig: ConfluenceConfiguration;
-  syncType?: UpsertDataSourceDocumentParams["upsertContext"]["sync_type"];
-  dataSourceConfig: DataSourceConfig;
-  loggerArgs: Record<string, string | number>;
-}
-
-async function upsertConfluencePageToDataSource({
-  page,
-  spaceName,
-  parents,
-  confluenceConfig,
-  syncType = "batch",
-  dataSourceConfig,
-  loggerArgs,
-}: ConfluenceUpsertPageInput) {
-  const localLogger = logger.child(loggerArgs);
-
-  const markdown = turndownService.turndown(page.body.storage.value);
-  const pageCreatedAt = new Date(page.createdAt);
-  const lastPageVersionCreatedAt = new Date(page.version.createdAt);
-
-  if (!markdown) {
-    logger.warn({ ...loggerArgs }, "Upserting page with empty content.");
-  }
-
-  const renderedMarkdown = await renderMarkdownSection(
-    dataSourceConfig,
-    markdown
-  );
-
-  // Log labels info
-  if (page.labels.results.length > 0) {
-    localLogger.info(
-      { labelsCount: page.labels.results.length },
-      "Confluence page has labels."
-    );
-  }
-
-  // Use label names for tags instead of IDs
-  const customTags = page.labels.results.map((l) => l.name);
-
-  const tags = [
-    `createdAt:${pageCreatedAt.getTime()}`,
-    `space:${spaceName}`,
-    `title:${page.title}`,
-    `updatedAt:${lastPageVersionCreatedAt.getTime()}`,
-    `version:${page.version.number}`,
-    ...filterCustomTags(customTags, localLogger),
-  ];
-
-  const renderedPage = await renderDocumentTitleAndContent({
-    dataSourceConfig,
-    title: `Page ${page.title}`,
-    createdAt: pageCreatedAt,
-    updatedAt: lastPageVersionCreatedAt,
-    content: renderedMarkdown,
-    additionalPrefixes: {
-      labels: page.labels.results.map((l) => l.name).join(", ") || "none",
-    },
-  });
-
-  const documentId = makePageInternalId(page.id);
-  const documentUrl = makeConfluenceDocumentUrl({
-    baseUrl: confluenceConfig.url,
-    suffix: page._links.tinyui,
-  });
-
-  await upsertDataSourceDocument({
-    dataSourceConfig,
-    documentContent: renderedPage,
-    documentId,
-    documentUrl,
-    loggerArgs,
-    parents,
-    parentId: parents[1],
-    tags,
-    timestampMs: lastPageVersionCreatedAt.getTime(),
-    upsertContext: { sync_type: syncType },
-    title: page.title,
-    mimeType: INTERNAL_MIME_TYPES.CONFLUENCE.PAGE,
-    async: true,
-  });
-}
-
-async function upsertConfluencePageInDb(
-  connectorId: ModelId,
-  page: ConfluencePageWithBodyType,
-  visitedAtMs: number
-) {
-  await ConfluencePage.upsert({
-    connectorId,
-    pageId: page.id,
-    spaceId: page.spaceId,
-    parentId: page.parentId,
-    parentType: page.parentType,
-    title: page.title,
-    externalUrl: page._links.tinyui,
-    version: page.version.number,
-    lastVisitedAt: new Date(visitedAtMs),
-  });
-}
-
 interface ConfluenceCheckAndUpsertSingleEntityActivityInput {
   connectorId: ModelId;
   entityRef: ConfluenceEntityRef;
@@ -440,7 +218,7 @@ export async function confluenceCheckAndUpsertSingleEntityActivity({
 
   switch (entityRef.type) {
     case "page":
-      return confluenceCheckAndUpsertSinglePageActivity({
+      return confluenceCheckAndUpsertSinglePage({
         connector,
         dataSourceConfig,
         pageRef: entityRef,
@@ -451,7 +229,7 @@ export async function confluenceCheckAndUpsertSingleEntityActivity({
       });
 
     case "folder":
-      return confluenceCheckAndUpsertSingleFolderActivity({
+      return confluenceCheckAndUpsertSingleFolder({
         connector,
         dataSourceConfig,
         folderRef: entityRef,
@@ -464,270 +242,6 @@ export async function confluenceCheckAndUpsertSingleEntityActivity({
     default:
       assertNever(entityRef);
   }
-}
-
-interface BaseConfluenceCheckAndUpsertSingleEntityActivityInput {
-  connector: ConnectorResource;
-  dataSourceConfig: DataSourceConfig;
-  forceUpsert: boolean;
-  isBatchSync: boolean;
-  space: SpaceBlob;
-  visitedAtMs: number;
-}
-
-/**
- * Upsert a Confluence page without its full parents.
- * Operates greedily by stopping if the page is restricted or if there is a version match
- * (unless the page was moved, in this case, we have to upsert because the parents have changed).
- */
-async function confluenceCheckAndUpsertSinglePageActivity({
-  connector,
-  dataSourceConfig,
-  pageRef,
-  forceUpsert,
-  isBatchSync,
-  space,
-  visitedAtMs,
-}: BaseConfluenceCheckAndUpsertSingleEntityActivityInput & {
-  pageRef: ConfluencePageRef;
-}) {
-  const { id: spaceId, name: spaceName } = space;
-  const { id: entityId } = pageRef;
-
-  const { id: connectorId } = connector;
-  const { id: pageId } = pageRef;
-
-  const loggerArgs = {
-    connectorId,
-    dataSourceId: dataSourceConfig.dataSourceId,
-    entityId,
-    spaceId,
-    workspaceId: dataSourceConfig.workspaceId,
-  };
-  const localLogger = logger.child(loggerArgs);
-
-  const pageAlreadyInDb = await ConfluencePage.findOne({
-    attributes: ["parentId", "skipReason", "version"],
-    where: {
-      connectorId,
-      pageId,
-    },
-  });
-
-  const isPageSkipped = Boolean(
-    pageAlreadyInDb && pageAlreadyInDb.skipReason !== null
-  );
-  if (isPageSkipped) {
-    logger.info("Confluence page skipped.");
-    return true;
-  }
-
-  const confluenceConfig =
-    await fetchConfluenceConfigurationActivity(connectorId);
-
-  const client = await getConfluenceClient(
-    {
-      cloudId: confluenceConfig.cloudId,
-    },
-    connector
-  );
-
-  // Check restrictions.
-  const { hasReadRestrictions } = pageRef;
-  if (hasReadRestrictions) {
-    localLogger.info("Skipping restricted Confluence page.");
-    return false;
-  }
-
-  // Check the version.
-  const isSameVersion =
-    pageAlreadyInDb && pageAlreadyInDb.version === pageRef.version;
-
-  // Check whether the page was moved (the version is not bumped when a page is moved).
-  const pageWasMoved =
-    pageAlreadyInDb && pageAlreadyInDb.parentId !== pageRef.parentId;
-
-  // Only index in DB if the page does not exist, has been moved, or we want to upsert.
-  if (isSameVersion && !forceUpsert && !pageWasMoved) {
-    // Simply record that we visited the page.
-    await markPageHasVisited({
-      connectorId,
-      pageId,
-      spaceId,
-      visitedAtMs,
-    });
-
-    return true;
-  }
-
-  // There is a small delta between the page being listed and the page being imported.
-  // If the page has been deleted in the meantime, we should ignore it.
-  const page = await client.getPageById(pageId);
-  if (!page) {
-    localLogger.info("Confluence page not found.");
-    // Return false to skip the child pages.
-    return false;
-  }
-
-  let parents: [string, string, ...string[]];
-  if (page.parentId && page.parentType) {
-    // Exact parent Ids will be computed after all page imports within the space have been completed.
-    parents = [
-      makePageInternalId(page.id),
-      makeEntityInternalId(page.parentType, page.parentId),
-      HiddenContentNodeParentId,
-    ];
-  } else {
-    // In this case we already have the exact parents: the page itself and the space.
-    parents = [makePageInternalId(page.id), makeSpaceInternalId(spaceId)];
-  }
-
-  localLogger.info("Upserting Confluence page.");
-  await upsertConfluencePageToDataSource({
-    page,
-    spaceName,
-    parents,
-    confluenceConfig,
-    syncType: isBatchSync ? "batch" : "incremental",
-    dataSourceConfig,
-    loggerArgs,
-  });
-
-  localLogger.info("Upserting Confluence page in DB.");
-  await upsertConfluencePageInDb(connector.id, page, visitedAtMs);
-
-  return true;
-}
-
-async function confluenceCheckAndUpsertSingleFolderActivity({
-  connector,
-  dataSourceConfig,
-  folderRef,
-  forceUpsert,
-  space,
-  visitedAtMs,
-}: BaseConfluenceCheckAndUpsertSingleEntityActivityInput & {
-  folderRef: ConfluenceFolderRef;
-}) {
-  const { id: spaceId } = space;
-  const { id: folderId } = folderRef;
-
-  const { id: connectorId } = connector;
-
-  const loggerArgs = {
-    connectorId,
-    dataSourceId: dataSourceConfig.dataSourceId,
-    folderId,
-    spaceId,
-    workspaceId: dataSourceConfig.workspaceId,
-  };
-  const localLogger = logger.child(loggerArgs);
-
-  const folderAlreadyInDb = await ConfluenceFolder.findOne({
-    attributes: ["parentId", "skipReason", "version"],
-    where: {
-      connectorId,
-      folderId,
-    },
-  });
-
-  const isFolderSkipped = Boolean(
-    folderAlreadyInDb && folderAlreadyInDb.skipReason !== null
-  );
-  if (isFolderSkipped) {
-    logger.info("Confluence folder skipped.");
-
-    // If a folder is skipped, we skip all its children.
-    return false;
-  }
-
-  const confluenceConfig =
-    await fetchConfluenceConfigurationActivity(connectorId);
-
-  const client = await getConfluenceClient(
-    {
-      cloudId: confluenceConfig.cloudId,
-    },
-    connector
-  );
-
-  // Check restrictions.
-  const { hasReadRestrictions } = folderRef;
-  if (hasReadRestrictions) {
-    localLogger.info("Skipping restricted Confluence folder.");
-    return false;
-  }
-
-  // Check the version.
-  const isSameVersion =
-    folderAlreadyInDb && folderAlreadyInDb.version === folderRef.version;
-
-  // Check whether the folder was moved (the version is not bumped when a folder is moved).
-  const folderWasMoved =
-    folderAlreadyInDb && folderAlreadyInDb.parentId !== folderRef.parentId;
-
-  // Only index in DB if the folder does not exist, has been moved, or we want to upsert.
-  if (isSameVersion && !forceUpsert && !folderWasMoved) {
-    // Simply record that we visited the folder.
-    await markFolderHasVisited({
-      connectorId,
-      folderId,
-      spaceId,
-      visitedAtMs,
-    });
-
-    return true;
-  }
-
-  // There is a small delta between the folder being listed and the folder being imported.
-  // If the folder has been deleted in the meantime, we should ignore it.
-  const folder = await client.getFolderById(folderId);
-  if (!folder) {
-    localLogger.info("Confluence folder not found.");
-    // Return false to skip the child entities.
-    return false;
-  }
-
-  let parents: [string, string, ...string[]];
-  if (folder.parentId && folder.parentType) {
-    // Exact parent Ids will be computed after all entity imports within the space have been completed.
-    parents = [
-      makeFolderInternalId(folder.id),
-      makeEntityInternalId(folder.parentType, folder.parentId),
-      HiddenContentNodeParentId,
-    ];
-  } else {
-    // In this case we already have the exact parents: the folder itself and the space.
-    parents = [makeFolderInternalId(folder.id), makeSpaceInternalId(spaceId)];
-  }
-
-  const parentId = parents[1];
-
-  localLogger.info("Upserting Confluence folder.");
-  await upsertDataSourceFolder({
-    dataSourceConfig: dataSourceConfigFromConnector(connector),
-    folderId: makeFolderInternalId(folderId),
-    mimeType: INTERNAL_MIME_TYPES.CONFLUENCE.FOLDER,
-    parentId,
-    parents,
-    sourceUrl: folder._links.tinyui,
-    title: folder.title,
-  });
-
-  localLogger.info("Upserting Confluence folder in DB.");
-  await ConfluenceFolder.upsert({
-    connectorId,
-    externalUrl: folder._links.tinyui,
-    folderId,
-    lastVisitedAt: new Date(visitedAtMs),
-    parentId: folder.parentId,
-    parentType: folder.parentType,
-    spaceId,
-    title: folder.title,
-    version: folder.version.number,
-  });
-
-  return true;
 }
 
 type ConfluenceUpsertLeafEntitiesActivityInput = Omit<
@@ -753,115 +267,6 @@ export async function confluenceUpsertLeafEntitiesActivity({
       concurrency: UPSERT_CONCURRENT_LIMIT,
     }
   );
-}
-
-/**
- * Upsert a Confluence page with its full parent hierarchy.
- * Expensive operation, it should be reserved to admin actions on a limited set of pages.
- */
-export async function confluenceUpsertPageWithFullParentsActivity({
-  connectorId,
-  pageId,
-  cachedSpaceNames = {},
-  cachedSpaceHierarchies = {},
-}: {
-  connectorId: ModelId;
-  pageId: string;
-  cachedSpaceNames?: Record<string, string>;
-  cachedSpaceHierarchies?: Record<
-    string,
-    Awaited<ReturnType<typeof getSpaceHierarchy>>
-  >;
-}): Promise<boolean> {
-  const connector = await ConnectorResource.fetchById(connectorId);
-  if (!connector) {
-    throw new Error("Connector not found.");
-  }
-  const dataSourceConfig = dataSourceConfigFromConnector(connector);
-  const confluenceConfig =
-    await fetchConfluenceConfigurationActivity(connectorId);
-
-  const loggerArgs = {
-    connectorId,
-    dataSourceId: dataSourceConfig.dataSourceId,
-    pageId,
-    workspaceId: dataSourceConfig.workspaceId,
-  };
-  const localLogger = logger.child(loggerArgs);
-  const visitedAtMs = new Date().getTime();
-
-  const pageInDb = await ConfluencePage.findOne({
-    attributes: ["parentId", "skipReason"],
-    where: { connectorId, pageId },
-  });
-  if (pageInDb && pageInDb.skipReason !== null) {
-    localLogger.info("Confluence page skipped.");
-    return false;
-  }
-
-  const client = await getConfluenceClient(
-    {
-      cloudId: confluenceConfig.cloudId,
-    },
-    connector
-  );
-
-  const hasReadRestrictions = await pageHasReadRestrictions(client, pageId);
-  if (hasReadRestrictions) {
-    localLogger.info("Skipping restricted Confluence page.");
-    return false;
-  }
-
-  const page = await client.getPageById(pageId);
-  if (!page) {
-    localLogger.info("Confluence page not found.");
-    return false;
-  }
-
-  let spaceName = cachedSpaceNames[page.spaceId];
-  if (!spaceName) {
-    const space = await client.getSpaceById(page.spaceId);
-    if (!space) {
-      localLogger.info("Confluence space not found.");
-      return false;
-    }
-    cachedSpaceNames[page.spaceId] = space.name;
-    spaceName = space.name;
-  }
-
-  if (!cachedSpaceHierarchies[page.spaceId]) {
-    cachedSpaceHierarchies[page.spaceId] = await getSpaceHierarchy(
-      connectorId,
-      page.spaceId
-    );
-  }
-
-  const parents = await getConfluenceEntityParentIds(
-    connectorId,
-    {
-      id: page.id,
-      parentId: page.parentId,
-      parentType: page.parentType,
-      spaceId: page.spaceId,
-      type: "page",
-    },
-    cachedSpaceHierarchies[page.spaceId]
-  );
-
-  localLogger.info("Upserting Confluence page.");
-  await upsertConfluencePageToDataSource({
-    page,
-    spaceName,
-    parents,
-    confluenceConfig,
-    dataSourceConfig,
-    loggerArgs,
-  });
-
-  localLogger.info("Upserting Confluence page in DB.");
-  await upsertConfluencePageInDb(connector.id, page, visitedAtMs);
-
-  return true;
 }
 
 export async function confluenceGetActiveChildEntityRefsActivity({
@@ -1037,7 +442,6 @@ export async function confluenceGetTopLevelEntityIdsActivity({
   };
 }
 
-// TODO:
 export async function confluenceUpdateEntitiesParentIdsActivity(
   connectorId: ModelId,
   spaceId: string,
@@ -1082,7 +486,7 @@ export async function confluenceUpdateEntitiesParentIdsActivity(
     "Start updating entities parent ids."
   );
 
-  // Utilize an in-memory map to cache entity hierarchies, thereby reducing database queries.
+  // Use an in-memory map to cache content hierarchies, thereby reducing database queries.
   const cachedHierarchy = await getSpaceHierarchy(connectorId, spaceId);
 
   await concurrentExecutor(
@@ -1090,9 +494,9 @@ export async function confluenceUpdateEntitiesParentIdsActivity(
     async (e) => {
       const isPage = e instanceof ConfluencePage;
 
-      // Retrieve parents using the internal ID, which aligns with the permissions
-      // view rendering and RAG requirements.
-      const parentIds = await getConfluenceEntityParentIds(
+      // Retrieve parents using the internal ID, which aligns with the permissions view rendering
+      // and RAG requirements.
+      const parentIds = await getConfluenceContentParentIds(
         connectorId,
         {
           id: isPage ? e.pageId : e.folderId,
@@ -1144,74 +548,20 @@ export async function confluenceRemoveUnvisitedEntitiesActivity({
   spaceId: string;
 }) {
   const connector = await fetchConfluenceConnector(connectorId);
-  // TODO: implement this for folders.
+  const dataSourceConfig = dataSourceConfigFromConnector(connector);
 
-  await confluenceRemoveUnvisitedPagesActivity({
+  await confluenceRemoveUnvisitedPages({
     connector,
+    dataSourceConfig,
     lastVisitedAt,
     spaceId,
   });
-}
 
-async function confluenceRemoveUnvisitedPagesActivity({
-  connector,
-  lastVisitedAt,
-  spaceId,
-}: {
-  connector: ConnectorResource;
-  lastVisitedAt: number;
-  spaceId: string;
-}) {
-  const { id: connectorId } = connector;
-
-  const unvisitedPages = await ConfluencePage.findAll({
-    attributes: ["pageId"],
-    where: {
-      connectorId,
-      spaceId,
-      lastVisitedAt: {
-        [Op.ne]: new Date(lastVisitedAt),
-      },
-    },
-  });
-
-  const dataSourceConfig = dataSourceConfigFromConnector(connector);
-
-  for (const page of unvisitedPages) {
-    // TODO(2024-01-22 flav) Add an extra check to ensure that the page does not exist anymore in Confluence.
-    await deletePage(connectorId, page.pageId, dataSourceConfig);
-  }
-}
-
-async function deletePage(
-  connectorId: ModelId,
-  pageId: string,
-  dataSourceConfig: DataSourceConfig
-) {
-  const loggerArgs = {
-    connectorId,
-    pageId,
-  };
-
-  const localLogger = logger.child(loggerArgs);
-
-  const documentId = makePageInternalId(pageId);
-  localLogger.info(
-    { documentId },
-    "Deleting Confluence page from Dust data source."
-  );
-
-  await deleteDataSourceDocument(dataSourceConfig, documentId, {
-    connectorId,
-    pageId,
-  });
-
-  localLogger.info("Deleting Confluence page from database.");
-  await ConfluencePage.destroy({
-    where: {
-      connectorId,
-      pageId,
-    },
+  await confluenceRemoveUnvisitedFolders({
+    connector,
+    dataSourceConfig,
+    lastVisitedAt,
+    spaceId,
   });
 }
 
@@ -1219,35 +569,23 @@ export async function confluenceRemoveSpaceActivity(
   connectorId: ModelId,
   spaceId: string
 ) {
-  const localLogger = logger.child({
-    spaceId,
-    connectorId,
-  });
-
   const connector = await fetchConfluenceConnector(connectorId);
 
   const dataSourceConfig = dataSourceConfigFromConnector(connector);
 
-  const allPages = await ConfluencePage.findAll({
-    attributes: ["pageId"],
-    where: {
-      connectorId,
-      spaceId,
-    },
+  await confluenceRemoveAllPagesInSpace({
+    connector,
+    dataSourceConfig,
+    spaceId,
   });
 
-  localLogger.info(
-    {
-      numberOfPages: allPages.length,
-    },
-    "Delete Confluence space"
-  );
+  await confluenceRemoveAllFoldersInSpace({
+    connector,
+    dataSourceConfig,
+    spaceId,
+  });
 
-  for (const page of allPages) {
-    await deletePage(connectorId, page.pageId, dataSourceConfig);
-  }
-
-  // deleting the folder in data_source_folders (core)
+  // Deleting the folder in data_source_folders (core).
   await deleteDataSourceFolder({
     dataSourceConfig,
     folderId: makeSpaceInternalId(spaceId),
@@ -1269,7 +607,122 @@ export async function fetchConfluenceSpaceIdsForConnectorActivity({
   return spacesForConnector.map((s) => s.spaceId);
 }
 
-// Personal Data Reporting logic.
+/**
+ * Page specific activities.
+ */
+
+/**
+ * Upsert a Confluence page with its full parent hierarchy.
+ * Expensive operation, it should be reserved to admin actions on a limited set of pages.
+ */
+export async function confluenceUpsertPageWithFullParentsActivity({
+  connectorId,
+  pageId,
+  cachedSpaceNames = {},
+  cachedSpaceHierarchies = {},
+}: {
+  connectorId: ModelId;
+  pageId: string;
+  cachedSpaceNames?: Record<string, string>;
+  cachedSpaceHierarchies?: Record<
+    string,
+    Awaited<ReturnType<typeof getSpaceHierarchy>>
+  >;
+}): Promise<boolean> {
+  const connector = await ConnectorResource.fetchById(connectorId);
+  if (!connector) {
+    throw new Error("Connector not found.");
+  }
+  const dataSourceConfig = dataSourceConfigFromConnector(connector);
+  const confluenceConfig =
+    await fetchConfluenceConfigurationActivity(connectorId);
+
+  const loggerArgs = {
+    connectorId,
+    dataSourceId: dataSourceConfig.dataSourceId,
+    pageId,
+    workspaceId: dataSourceConfig.workspaceId,
+  };
+  const localLogger = logger.child(loggerArgs);
+  const visitedAtMs = new Date().getTime();
+
+  const pageInDb = await ConfluencePage.findOne({
+    attributes: ["parentId", "skipReason"],
+    where: { connectorId, pageId },
+  });
+  if (pageInDb && pageInDb.skipReason !== null) {
+    localLogger.info("Confluence page skipped.");
+    return false;
+  }
+
+  const client = await getConfluenceClient(
+    {
+      cloudId: confluenceConfig.cloudId,
+    },
+    connector
+  );
+
+  const hasReadRestrictions = await pageHasReadRestrictions(client, pageId);
+  if (hasReadRestrictions) {
+    localLogger.info("Skipping restricted Confluence page.");
+    return false;
+  }
+
+  const page = await client.getPageById(pageId);
+  if (!page) {
+    localLogger.info("Confluence page not found.");
+    return false;
+  }
+
+  let spaceName = cachedSpaceNames[page.spaceId];
+  if (!spaceName) {
+    const space = await client.getSpaceById(page.spaceId);
+    if (!space) {
+      localLogger.info("Confluence space not found.");
+      return false;
+    }
+    cachedSpaceNames[page.spaceId] = space.name;
+    spaceName = space.name;
+  }
+
+  if (!cachedSpaceHierarchies[page.spaceId]) {
+    cachedSpaceHierarchies[page.spaceId] = await getSpaceHierarchy(
+      connectorId,
+      page.spaceId
+    );
+  }
+
+  const parents = await getConfluenceContentParentIds(
+    connectorId,
+    {
+      id: page.id,
+      parentId: page.parentId,
+      parentType: page.parentType,
+      spaceId: page.spaceId,
+      type: "page",
+    },
+    cachedSpaceHierarchies[page.spaceId]
+  );
+
+  localLogger.info("Upserting Confluence page.");
+  await confluenceUpsertPageToDataSource({
+    page,
+    spaceName,
+    parents,
+    confluenceConfig,
+    dataSourceConfig,
+    loggerArgs,
+  });
+
+  localLogger.info("Upserting Confluence page in DB.");
+  await upsertConfluencePageInDb(connector.id, page, visitedAtMs);
+
+  return true;
+}
+
+/**
+ * Personal Data Reporting logic.
+ */
 
 interface ConfluenceUserAccountAndConnectorId {
   connectorId: ModelId;
