+++ front/lib/api/files/client_executable.ts
@@ -125,6 +125,7 @@ export async function editClientExecutableFile(
   }
 
   // Perform the replacement.
+  // FIXME: This reset the shareable file record.
   const updatedContent = currentContent.replace(regex, newString);
 
   // Upload the updated content.
